+++ tests/unit_tests/test_merger.py
@@ -37,11 +37,11 @@ def test_create_merged_line(merger: LineMerger) -> None:
     ).parse_query(source_string)
 
     expected = "select able, baker,\n"
-    actual = merger.create_merged_line(raw_query.lines[0:4])
+    actual = merger.create_merged_line(raw_query.lines[0:4])[0]
     assert str(actual) == expected
 
     expected = "select able, baker, charlie, delta,\n"
-    actual = merger.create_merged_line(raw_query.lines)
+    actual = merger.create_merged_line(raw_query.lines)[0]
     assert str(actual) == expected
 
     with pytest.raises(CannotMergeException):
@@ -257,8 +257,8 @@ def test_merge_single_line(merger: LineMerger) -> None:
     q = merger.mode.dialect.initialize_analyzer(merger.mode.line_length).parse_query(
         source_string
     )
-    merged_line = merger.create_merged_line(q.lines)
-    assert merged_line == q.lines[0]
+    merged_lines = merger.create_merged_line(q.lines)
+    assert merged_lines == q.lines
 
 
 def test_merge_lines_split_by_operators(merger: LineMerger) -> None:
@@ -306,3 +306,34 @@ def test_do_not_merge_very_long_chains(merger: LineMerger) -> None:
     result_string = "".join([str(line) for line in merged_lines])
 
     assert result_string == source_string
+
+
+def test_respect_extra_blank_lines(merger: LineMerger) -> None:
+    source_string = """
+    select
+        one_field,
+        another_field,
+        yet_another_field
+
+
+    from
+        my_table
+
+    join
+        your_table
+        on my_table.your_id
+        = your_table.my_id
+
+
+    where
+        something is true
+    """.lstrip()
+    raw_query = merger.mode.dialect.initialize_analyzer(
+        merger.mode.line_length
+    ).parse_query(source_string)
+    merged_lines = merger.maybe_merge_lines(raw_query.lines)
+    assert merged_lines[1].is_blank_line
+    assert merged_lines[2].is_blank_line
+    assert merged_lines[4].is_blank_line
+    assert merged_lines[6].is_blank_line
+    assert merged_lines[7].is_blank_line